Extract expression module
